[ANDROID] Fix android after merge
