+++ android/build.gradle
@@ -45,5 +45,7 @@ dependencies {
     implementation('com.google.android.exoplayer:extension-okhttp:2.18.1') {
         exclude group: 'com.squareup.okhttp3', module: 'okhttp'
     }
+    implementation 'com.google.android.exoplayer:extension-ima:2.11.4'
+
     implementation "com.squareup.okhttp3:okhttp:" + '$OKHTTP_VERSION'
 }
