+++ android/src/main/java/com/brentvatne/exoplayer/ExoPlayerView.java
@@ -19,16 +19,19 @@ import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.Timeline;
+import com.google.android.exoplayer2.source.ads.AdsLoader;
 import com.google.android.exoplayer2.Tracks;
 import com.google.android.exoplayer2.text.Cue;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.ui.SubtitleView;
+import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.video.VideoSize;
 
 import java.util.List;
+import java.util.ArrayList;
 
 @TargetApi(16)
-public final class ExoPlayerView extends FrameLayout {
+public final class ExoPlayerView extends FrameLayout implements AdsLoader.AdViewProvider {
 
     private View surfaceView;
     private final View shutterView;
@@ -38,6 +41,7 @@ public final class ExoPlayerView extends FrameLayout {
     private ExoPlayer player;
     private Context context;
     private ViewGroup.LayoutParams layoutParams;
+    private final FrameLayout adOverlayFrameLayout;
 
     private boolean useTextureView = true;
     private boolean useSecureView = false;
@@ -83,7 +87,10 @@ public final class ExoPlayerView extends FrameLayout {
         layout.addView(shutterView, 1, layoutParams);
         layout.addView(subtitleLayout, 2, layoutParams);
 
+        adOverlayFrameLayout = new FrameLayout(context);
+
         addViewInLayout(layout, 0, aspectRatioParams);
+        addViewInLayout(adOverlayFrameLayout, 1, layoutParams);
     }
 
     private void clearVideoView() {
@@ -137,6 +144,28 @@ public final class ExoPlayerView extends FrameLayout {
 
     private void updateShutterViewVisibility() {
         shutterView.setVisibility(this.hideShutterView ? View.INVISIBLE : View.VISIBLE);
+    }
+
+    @Override
+    public void requestLayout() {
+        super.requestLayout();
+        post(measureAndLayout);
+    }
+
+     // AdsLoader.AdViewProvider implementation.
+
+    @Override
+    public ViewGroup getAdViewGroup() {
+        return Assertions.checkNotNull(adOverlayFrameLayout, "exo_ad_overlay must be present for ad playback");
+    }
+
+    @Override
+    public View[] getAdOverlayViews() {
+        ArrayList<View> overlayViews = new ArrayList<>();
+        if (adOverlayFrameLayout != null) {
+            overlayViews.add(adOverlayFrameLayout);
+        }
+        return overlayViews.toArray(new View[0]);
     }
 
     /**
